requirements.txt
flask==3.1.3
flask-cors==6.0.2
scikit-learn==1.8.0
numpy==2.4.3
pymysql==1.1.2
python-dotenv==1.2.2
gunicorn==23.0.0